Integrated cache for CatalogService. Created cache for pagination of the first 100 clothing items and the first 30 items for the stock.

+++ Clothy.Services/Clothy.CatalogService/Clothy.CatalogService.BLL/Services/ClotheService.cs
@@ -9,12 +9,8 @@ using Clothy.CatalogService.Domain.QueryParameters;
 using Clothy.Shared.Exceptions;
 using Clothy.Shared.Helpers;
 using Clothy.Shared.Helpers.CloudinaryConfig;
+using Clothy.Shared.Cache.Interfaces;
 using Microsoft.AspNetCore.Http;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace Clothy.CatalogService.BLL.Services
 {
@@ -23,71 +19,98 @@ namespace Clothy.CatalogService.BLL.Services
         private IUnitOfWork unitOfWork;
         private IMapper mapper;
         private IImageService imageService;
+        private IEntityCacheService cacheService;
+        private IEntityCacheInvalidationService<ClotheItem> cacheInvalidationService;
+        private const int MAX_CASHED_PAGES = 10;
 
-        public ClotheService(IUnitOfWork unitOfWork, IMapper mapper, IImageService imageService)
+        public ClotheService(IUnitOfWork unitOfWork,IMapper mapper, IImageService imageService, IEntityCacheService cacheService, IEntityCacheInvalidationService<ClotheItem> cacheInvalidationService)
         {
             this.unitOfWork = unitOfWork;
             this.mapper = mapper;
             this.imageService = imageService;
+            this.cacheService = cacheService;
+            this.cacheInvalidationService = cacheInvalidationService;
         }
 
         public async Task<PagedList<ClotheSummaryDTO>> GetPagedClotheItemsAsync(ClotheItemSpecificationParameters parameters, CancellationToken cancellationToken = default)
         {
-            PagedList<ClotheItem> paged = await unitOfWork.ClotheItems.GetPagedClotheItemsAsync(parameters, cancellationToken);
-            List<ClotheSummaryDTO> mapped = mapper.Map<List<ClotheSummaryDTO>>(paged.Items);
-            
-            return new PagedList<ClotheSummaryDTO>(mapped, paged.TotalCount, paged.CurrentPage, paged.PageSize);
+            bool usePageCache = parameters.PageNumber <= MAX_CASHED_PAGES;
+            string cacheKey;
+
+            if (usePageCache) cacheKey = $"clothes:page:{parameters.PageNumber}:size:{parameters.PageSize}";
+            else
+            {
+                PagedList<ClotheItem> paged = await unitOfWork.ClotheItems.GetPagedClotheItemsAsync(parameters, cancellationToken);
+                List<ClotheSummaryDTO> mapped = mapper.Map<List<ClotheSummaryDTO>>(paged.Items);
+                return new PagedList<ClotheSummaryDTO>(mapped, paged.TotalCount, paged.CurrentPage, paged.PageSize);
+            }
+
+            var cached = await cacheService.GetOrSetAsync(cacheKey, async () =>
+            {
+                PagedList<ClotheItem> paged = await unitOfWork.ClotheItems.GetPagedClotheItemsAsync(parameters, cancellationToken);
+                List<ClotheSummaryDTO> mapped = mapper.Map<List<ClotheSummaryDTO>>(paged.Items);
+                return new PagedList<ClotheSummaryDTO>(mapped, paged.TotalCount, paged.CurrentPage, paged.PageSize);
+            });
+
+            return cached!;
         }
+
 
         public async Task<ClotheDetailDTO> GetDetailByIdAsync(Guid id, CancellationToken cancellationToken = default)
         {
-            ClotheItem? clotheItem = await unitOfWork.ClotheItems.GetByIdWithDetailsAsync(id, cancellationToken);
-            if (clotheItem == null) throw new NotFoundException($"Clothe item not found with ID: {id}");
-            
-            return mapper.Map<ClotheDetailDTO>(clotheItem);
+            string cacheKey = $"clothe:{id}";
+            var cached = await cacheService.GetOrSetAsync(cacheKey, async () =>
+            {
+                ClotheItem? clotheItem = await unitOfWork.ClotheItems.GetByIdWithDetailsAsync(id, cancellationToken);
+                if (clotheItem == null) throw new NotFoundException($"Clothe item not found with ID: {id}");
+                
+                return mapper.Map<ClotheDetailDTO>(clotheItem);
+            });
+
+            return cached!;
         }
 
         public async Task<ClotheDetailDTO> CreateAsync(ClotheCreateDTO dto, CancellationToken cancellationToken = default)
         {
             if (await unitOfWork.ClotheItems.IsSlugAlreadyExistsAsync(dto.Slug, null, cancellationToken)) throw new AlreadyExistsException("Clothe with this slug already exists");
 
-            int totalPercentage = dto.Materials.Sum(percentage => percentage.Percentage);
+            int totalPercentage = dto.Materials.Sum(p => p.Percentage);
             if (totalPercentage != 100) throw new InvalidMaterialPercentageException("Total material percentage must be exactly 100.");
 
             ClotheItem clothe = mapper.Map<ClotheItem>(dto);
-
             clothe.MainPhotoURL = await imageService.UploadAsync(dto.MainPhoto, "clothes");
 
             clothe.Photos = new List<PhotoClothes>();
-            foreach (IFormFile photo in dto.AdditionalPhotos)
+            foreach (var photo in dto.AdditionalPhotos)
             {
                 string url = await imageService.UploadAsync(photo, "clothes");
-                clothe.Photos.Add(new PhotoClothes { 
+                clothe.Photos.Add(new PhotoClothes 
+                { 
                     PhotoURL = url 
                 });
             }
 
-            if (!await unitOfWork.Materials.AreAllExistAsync(dto.Materials.Select(m => m.MaterialId), cancellationToken)) throw new NotFoundException("One or more materials do not exist.");
-
+            if (!await unitOfWork.Materials.AreAllExistAsync(dto.Materials.Select(material => material.MaterialId), cancellationToken)) throw new NotFoundException("One or more materials do not exist.");
             if (!await unitOfWork.Tags.AreAllExistAsync(dto.TagIds, cancellationToken)) throw new NotFoundException("One or more tags do not exist.");
 
             clothe.ClotheMaterials = dto.Materials
-                .Select(material => new ClotheMaterial 
+                .Select(clotheMaterial => new ClotheMaterial 
                 { 
-                    MaterialId = material.MaterialId, 
-                    Percentage = material.Percentage 
+                    MaterialId = clotheMaterial.MaterialId, 
+                    Percentage = clotheMaterial.Percentage 
                 })
                 .ToList();
 
             clothe.ClotheTags = dto.TagIds
-                .Select(clotheTag => new ClotheTag 
-                { 
-                    TagId = clotheTag 
+                .Select(tagId => new ClotheTag { 
+                    TagId = tagId 
                 })
                 .ToList();
 
             await unitOfWork.ClotheItems.AddAsync(clothe, cancellationToken);
             await unitOfWork.SaveChangesAsync(cancellationToken);
+
+            await cacheInvalidationService.InvalidateAllAsync();
 
             return await GetDetailByIdAsync(clothe.Id, cancellationToken);
         }
@@ -106,10 +129,7 @@ namespace Clothy.CatalogService.BLL.Services
 
             if (dto.MainPhoto != null)
             {
-                if (!string.IsNullOrEmpty(clotheItem.MainPhotoURL))
-                {
-                    await imageService.DeleteImageAsync(clotheItem.MainPhotoURL);
-                }
+                if (!string.IsNullOrEmpty(clotheItem.MainPhotoURL)) await imageService.DeleteImageAsync(clotheItem.MainPhotoURL);
                 clotheItem.MainPhotoURL = await imageService.UploadAsync(dto.MainPhoto, "clothes");
             }
 
@@ -118,32 +138,28 @@ namespace Clothy.CatalogService.BLL.Services
                 foreach (IFormFile photo in dto.AdditionalPhotos)
                 {
                     string url = await imageService.UploadAsync(photo, "clothes");
-                    clotheItem.Photos.Add(new PhotoClothes 
-                    { 
+                    clotheItem.Photos.Add(new PhotoClothes { 
                         PhotoURL = url 
                     });
                 }
             }
 
-            if (!await unitOfWork.Materials.AreAllExistAsync(dto.Materials.Select(m => m.MaterialId), cancellationToken)) throw new NotFoundException("One or more materials do not exist.");
-
+            if (!await unitOfWork.Materials.AreAllExistAsync(dto.Materials.Select(material => material.MaterialId), cancellationToken)) throw new NotFoundException("One or more materials do not exist.");
             if (!await unitOfWork.Tags.AreAllExistAsync(dto.TagIds, cancellationToken)) throw new NotFoundException("One or more tags do not exist.");
 
             clotheItem.ClotheMaterials.Clear();
-            foreach (ClotheMaterialCreateDTO materialDto in dto.Materials)
+            foreach (ClotheMaterialCreateDTO material in dto.Materials)
             {
-                clotheItem.ClotheMaterials.Add(new ClotheMaterial 
-                { 
-                    MaterialId = materialDto.MaterialId, 
-                    Percentage = materialDto.Percentage 
+                clotheItem.ClotheMaterials.Add(new ClotheMaterial { 
+                    MaterialId = material.MaterialId, 
+                    Percentage = material.Percentage 
                 });
             }
 
             clotheItem.ClotheTags.Clear();
             foreach (Guid tagId in dto.TagIds)
             {
-                clotheItem.ClotheTags.Add(new ClotheTag 
-                { 
+                clotheItem.ClotheTags.Add(new ClotheTag { 
                     TagId = tagId 
                 });
             }
@@ -151,17 +167,18 @@ namespace Clothy.CatalogService.BLL.Services
             unitOfWork.ClotheItems.Update(clotheItem);
             await unitOfWork.SaveChangesAsync(cancellationToken);
 
+            await cacheInvalidationService.InvalidateByIdAsync(id);
+
             return await GetDetailByIdAsync(clotheItem.Id, cancellationToken);
         }
 
         public async Task<PriceRangeDTO> GetMinAndMaxPriceAsync(CancellationToken cancellationToken = default)
         {
-            var (minPrice, maxPrice) = await unitOfWork.ClotheItems.GetMinAndMaxPriceAsync(cancellationToken);
-
-            return new PriceRangeDTO
-            {
-                MinPrice = minPrice,
-                MaxPrice = maxPrice
+            (decimal minPrice, decimal maxPrice) = await unitOfWork.ClotheItems.GetMinAndMaxPriceAsync(cancellationToken);
+           
+            return new PriceRangeDTO { 
+                MinPrice = minPrice, 
+                MaxPrice = maxPrice 
             };
         }
 
@@ -170,21 +187,16 @@ namespace Clothy.CatalogService.BLL.Services
             ClotheItem? clothe = await unitOfWork.ClotheItems.GetByIdWithDetailsAsync(id, cancellationToken);
             if (clothe == null) throw new NotFoundException($"Clothe item not found with ID: {id}");
 
-            if (!string.IsNullOrEmpty(clothe.MainPhotoURL))
-            {
-                await imageService.DeleteImageAsync(clothe.MainPhotoURL);
-            }
+            if (!string.IsNullOrEmpty(clothe.MainPhotoURL)) await imageService.DeleteImageAsync(clothe.MainPhotoURL);
 
             foreach (PhotoClothes photo in clothe.Photos)
             {
-                if (!string.IsNullOrEmpty(photo.PhotoURL))
-                {
-                    await imageService.DeleteImageAsync(photo.PhotoURL);
-                }
+                if (!string.IsNullOrEmpty(photo.PhotoURL)) await imageService.DeleteImageAsync(photo.PhotoURL);
             }
 
             unitOfWork.ClotheItems.Delete(clothe);
             await unitOfWork.SaveChangesAsync(cancellationToken);
+            await cacheInvalidationService.InvalidateByIdAsync(id);
         }
     }
 }
